pyproject.toml
[project]
name = "myagent"
version = "0.1.0"
description = "AI agents built with DeepAgents framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.3.5",
    "python-dotenv>=1.0.0",
    "langchain-openai>=1.1.7",
    "langchain-anthropic>=0.3.0",
    "tavily-python>=0.7.18",
    "httpx>=0.27.0",
    "markdownify>=0.14.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
langgraph = [
    "langgraph-cli>=0.1.0",
]
